+++ mud-contracts/world/src/codegen/index.sol
@@ -25,9 +25,12 @@ import { EphemeralInvTable, EphemeralInvTableData } from "./tables/EphemeralInvT
 import { EphemeralInvItemTable, EphemeralInvItemTableData } from "./tables/EphemeralInvItemTable.sol";
 import { ItemTransferOffchainTable, ItemTransferOffchainTableData } from "./tables/ItemTransferOffchainTable.sol";
 import { SmartTurretConfigTable } from "./tables/SmartTurretConfigTable.sol";
+import { SmartGateConfigTable, SmartGateConfigTableData } from "./tables/SmartGateConfigTable.sol";
+import { SmartGateLinkTable } from "./tables/SmartGateLinkTable.sol";
 import { Balances } from "./tables/Balances.sol";
 import { TokenURI } from "./tables/TokenURI.sol";
 import { Owners } from "./tables/Owners.sol";
 import { ERC721Registry } from "./tables/ERC721Registry.sol";
 import { TokenApproval } from "./tables/TokenApproval.sol";
 import { OperatorApproval } from "./tables/OperatorApproval.sol";
+import { KillMailTable, KillMailTableData } from "./tables/KillMailTable.sol";