+++ mud-contracts/world/src/modules/kill-mail/Utils.sol
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: MIT
+pragma solidity >=0.8.21;
+
+import { ResourceId, WorldResourceIdLib, WorldResourceIdInstance } from "@latticexyz/world/src/WorldResourceId.sol";
+import { RESOURCE_SYSTEM, RESOURCE_TABLE } from "@latticexyz/world/src/worldResourceTypes.sol";
+import "./constants.sol";
+
+library Utils {
+  using WorldResourceIdInstance for ResourceId;
+
+  function killMailTableId(bytes14 namespace) internal pure returns (ResourceId) {
+    return WorldResourceIdLib.encode({ typeId: RESOURCE_TABLE, namespace: namespace, name: KILL_MAIL_TABLE_NAME });
+  }
+
+  function killMailSystemId(bytes14 namespace) internal pure returns (ResourceId) {
+    return WorldResourceIdLib.encode({ typeId: RESOURCE_SYSTEM, namespace: namespace, name: KILL_MAIL_SYSTEM_NAME });
+  }
+}