+++ mud-contracts/world/CHANGELOG.md
@@ -1,4 +1,16 @@
 # @eveworld/world
+## 0.0.12
+- Including Gate and Kill Mail in mud.config to generate abis 
+- Adding scripts for end-to-end tests 
+
+## 0.0.11
+
+### Patch Changes
+
+- smart turret feature implementation
+- update smart character creation (Breaking Change in interface)
+  - added corpId to `createCharacter` function
+  - added function to update corpId
 
 ## 0.0.11